inputs val change -> signal slot -> update display widgets

+++ global.h
@@ -100,7 +100,7 @@ public:
 
     struct  inOut{
         bool Di = false;
-        float An = 0.0;
+        int An = 0.0;
         QString name = "IN/OUT";
     } ;
 
@@ -143,8 +143,7 @@ public:
 
 
     QColor  backgroundColor = QColor(250, 250, 250, 175);   // 255 transparent
-    //WidgetDiagramElement * ptrCurrWidget = nullptr;
-     //       comPortptr = nullptr
+
 private:
 
     void create_IN_OUT_list();
